Remove all references to Swing (and by extension AWT).
As a bonus, this means all the `synchronized` modifiers can be removed
since all the methods are called on the same thread. This does not
appear to have any noticeable impact on the game, but theoretically
improves performance.

More importantly, this dodges potentially initializing AWT's toolkit,
which could otherwise cause crashes for Mac users.